Trabalho definitivamente pronto. Testar manualmente agora

+++ Rubro_negro/Includes/Objetos/CEP.h
@@ -10,7 +10,7 @@ void liberar_CEP(char **cep);
 
 void corrigir_formatacao_CEP(char **cep);
 
-int validar_CEP(char *cep);
+short int validar_CEP(char *cep);
 
 char *digitar_CEP();
