+++ Rubro_negro/Includes/Objetos/CPF.h
@@ -5,6 +5,6 @@ void imprimir_cpf(char *cpf);
 
 void corrigir_formatacao_cpf(char **cpf);
 
-int validar_cpf(char *cpf);
+short int validar_cpf(char *cpf);
 
 char *digitar_CPF();
